CAmbio de logiaca para que funcione la api en frond, antes funcionaba en servidor
