+++ app-clima-matias-torres/components/InputCiudad.js
@@ -1,18 +1,33 @@
 "use client";
 
 import React, { useState } from "react";
+let clave = process.env.CLAVE_API;
 
 const InputCiudad = () => {
   const [city, setCity] = useState("");
   const [error, setError] = useState("");
 
-  const handleSubmit = (event) => {
+  var requestOptions = {
+    method: "GET",
+    redirect: "follow",
+  };
+
+  const handleSubmit = async (event) => {
     event.preventDefault();
     if (city.trim() === "") {
       setError("Por favor, ingresa una ciudad");
     } else {
       setError("");
       console.log("Ciudad ingresada:", city);
+      const data = await fetch(
+        `https://api.openweathermap.org/data/2.5/weather?q=${city}&appid=5f3da5be43514d2f2fa95b7d0e3e01d6`,
+        requestOptions
+      )
+        .then((response) => response.text())
+        .then((result) => console.log(result))
+        .catch((error) => console.log("error", error));
+
+      console.log(data);
     }
   };
