+++ app-clima-matias-torres/pages/index.js
@@ -4,14 +4,10 @@ import { Inter } from "next/font/google";
 import styles from "@/styles/Home.module.css";
 import InputCiudad from "@/components/InputCiudad";
 import { useState } from "react";
-import FetchClimate from "@/utils";
 
-const apiKey = process.env.CLIMA_API;
 const inter = Inter({ subsets: ["latin"] });
 
 export default function Home() {
-  const [data, setData] = useState(null);
-
   return (
     <>
       <Head>
